refactor: move non-generic ui components out of ui folder
