+++ zenith/src/lib/components/launch-list/launch-item.svelte
